feat(core/ui): widgets -> layouts
