+++ core/src/trezor/ui/components/common/__init__.py
@@ -2,29 +2,10 @@
 The model/common module contains code that is used by both model/tt and model/t1.
 """
 
-from trezor import log
-from trezor.messages import ButtonRequestType
-
-from apps.common import button_request
-
 if False:
-    from typing import List, Union, Awaitable, Any
+    from typing import List, Union
 
     TextContent = Union[str, int]
-
-    from trezor import wire
-    from trezor.messages.ButtonRequest import EnumTypeButtonRequestType
-
-
-async def interact(
-    ctx: wire.GenericContext,
-    widget: Awaitable,
-    brtype: str,
-    brcode: EnumTypeButtonRequestType = ButtonRequestType.Other,
-) -> Any:
-    log.debug(__name__, "ButtonRequest.type={}".format(brtype))
-    await button_request(ctx, code=brcode)
-    return await ctx.wait(widget)
 
 
 def break_path_to_lines(path_str: str, per_line: int) -> List[str]:
